[Added] TypeScript Support
With this we are moving the extension to TypeScript.
