+++ package.json
@@ -20,7 +20,7 @@
         "onCommand:extension.compressFolder",
         "onCommand:extension.getCompressionCount"
     ],
-    "main": "./extension",
+    "main": "./out/extension.js",
     "contributes": {
         "configuration": [
             {
@@ -107,17 +107,21 @@
     },
     "scripts": {
         "postinstall": "node ./node_modules/vscode/bin/install",
-        "test": "node ./node_modules/vscode/bin/test"
+        "test": "node ./node_modules/vscode/bin/test",
+        "vscode:prepublish": "npm run compile",
+        "compile": "tsc -p tsconfig.json",
+        "watch": "tsc -watch -p tsconfig.json"
     },
     "devDependencies": {
         "@types/mocha": "^8.2.1",
         "@types/node": "^14.14.35",
+        "@types/vscode": "^1.54.0",
         "eslint": "^7.22.0",
         "typescript": "^4.2.3",
         "vscode": "^1.1.37"
     },
     "dependencies": {
-        "tinify": "^1.5.0"
+        "tinify": "^1.6.0-beta.2"
     },
     "bugs": {
         "url": "https://github.com/andi1984/vscode-tinypng/issues",
