+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es2019",
+        "lib": ["ES2019"],
+        "outDir": "out",
+        "sourceMap": true,
+        "skipLibCheck": true,
+        /* Strict Type-Checking Option */
+        "strict": true /* enable all strict type-checking options */,
+        /* Additional Checks */
+        "noUnusedLocals": true /* Report errors on unused locals. */
+        // "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
+        // "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
+        // "noUnusedParameters": true,  /* Report errors on unused parameters. */
+    },
+    "include": ["./extension.ts"],
+    "exclude": ["node_modules"]
+}
